MNT #140 relocate functions per review
